Move client-window creation into the v2 runtime
Centralize presentation mode selection so integrations reuse one runtime-owned factory.
